+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Updated readme.
 - Install dragon on `npm i`.
+- Override CI checks if not building self.
 
 ## [2.0.2] - 2019-30-11